DP_RM_CVAR_ALTERTYPE and DP_RM_CVAR_WATCHED extensions/
DP_RM_CVAR_WATCHED: allows qc progs to track cvar changes. If present, the CvarUpdated function is called whenever a watched cvar changes its value. The first argument is the name of the cvar, the second is the previous string value. The current value can be retrieved using cvar_string or the associated autocvar. This function shouldn't return anything.

DP_RM_CVAR_ALTERTYPE: allows qc progs to change some of the cvar flags, using a new builtin function:

    float(string varname, float setflags, float unsetflags) cvar_altertype = #505;

It uses the same flags as cvar_type, not the actual cvar flags that are used internally. Currently, only CVAR_TYPEFLAG_SAVED, CVAR_TYPEFLAG_READONLY and CVAR_TYPEFLAG_WATCHED are allowed to be changed, and it can't be used on engine cvars. Returns 0 on error (in this case the cvar is not modified, even if some of the passed flags are correct), 1 otherwise.