Very basic IME support (SDL2 only), so you can brag about your 素晴らしい凄い可愛い日本語です or whatever else requires indirect input, right in the game chat.
Tested only with IBus+Anthy on Linux so far. You will need SDL2 compiled with IBus support for this to work, and not all (if any?) distributions provide it out of the box. I had to compile it myself on Arch Linux at least.
It should work with any other IME that SDL supports on your system.